supply-chain-agent/requirements.txt
# Agent runtime deps (used by bedrock-agentcore-starter-toolkit container build)
a2a-sdk>=0.3.0
click>=8.1.8
dotenv>=0.9.9
httpx>=0.28.1
langchain-google-genai>=2.1.4
langgraph>=0.4.1
pydantic>=2.11.4
python-dotenv>=1.1.0
sse-starlette>=2.3.5
starlette>=0.46.2
uvicorn>=0.34.2
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-httpx>=0.40b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# AgentCore CLI (for agentcore configure, dev, deploy, invoke)
bedrock-agentcore-starter-toolkit
boto3>=1.34.0
PyYAML>=6.0
